Render automations on home
