+++ crosscompute/macros/__init__.py
@@ -2,3 +2,4 @@
 from .disk import make_folder
 from .log import format_path
 from .text import compact_whitespace, normalize_key
+from .web import get_slug_from_name
